Changed version number from 1.8.2 to 1.8.3
git-svn-id: http://svn.code.sf.net/p/seb/svn/trunk@1412 0c343a18-9ee7-46a8-87ea-1d1407b43e6f

+++ win/SebWindowsPackage/SebWindowsService/ProjectInstaller.cs
@@ -86,7 +86,7 @@ namespace SebWindowsService
 
             string Manufacturer = "ETH Zuerich";
             string Product      = "SEB Windows";
-            string Version      = "1.8.2";
+            string Version      = "1.8.3";
             string Component    = "SebWindowsClient";
             string Build        = "Release";
 
@@ -120,7 +120,7 @@ namespace SebWindowsService
 
             Boolean ShortcutDesktop = (SebShortcut != string.Empty);
 
-
+/*
             // Write some debug data into a file
             string UserDesktopDir = Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory);
             string UserDebugFile  = UserDesktopDir + "\\" + "ContextParameters.txt";
@@ -153,7 +153,7 @@ namespace SebWindowsService
                 sw.WriteLine();
                 sw.Flush();
             }
-
+*/
 
             string SebBatchDir   = SebSourceDir;
             string SebConfigDir  = ProgramData   + "\\" + Manufacturer + "\\" + Product + " " + Version;
@@ -360,7 +360,7 @@ namespace SebWindowsService
 
             string Manufacturer = "ETH Zuerich";
             string Product      = "SEB Windows";
-            string Version      = "1.8.2";
+            string Version      = "1.8.3";
             string Component    = "SebWindowsClient";
             string Build        = "Release";
 
@@ -387,11 +387,11 @@ namespace SebWindowsService
             // ATTENTION:
             //
             // Deleting the SEB configuration directory in the ProgramData directory, e.g.
-            // C:\ProgramData\ETH Zuerich\SEB Windows 1.8.2 ,
+            // C:\ProgramData\ETH Zuerich\SEB Windows 1.8.3 ,
             // mostly succeeds.
             //
             // Deleting the SEB installation directory in the Program Files directory, e.g.
-            // C:\Program Files (x86)\ETH Zuerich\SEB Windows 1.8.2\SebWindowsClient\Release ,
+            // C:\Program Files (x86)\ETH Zuerich\SEB Windows 1.8.3\SebWindowsClient\Release ,
             // mostly fails, even though its files have all been deleted before.
             //
             // This is a known and annoying Windows bug still occurring in Windows 7:
@@ -400,7 +400,7 @@ namespace SebWindowsService
             //
             // Currently, only a user logoff or machine reboot solves this,
             // so maybe it is necessary to reboot and manually delete the
-            // C:\Program Files (x86)\ETH Zuerich\SEB Windows 1.8.2\SebWindowsClient\Release
+            // C:\Program Files (x86)\ETH Zuerich\SEB Windows 1.8.3\SebWindowsClient\Release
             // directory after reboot.
 
 
@@ -441,7 +441,8 @@ namespace SebWindowsService
 
 
             // Try to delete the "SebWindowsService.logfile.txt" file
-            System.IO.File.Delete(SebServiceLog);
+            if (System.IO.File.Exists(SebServiceLog))
+                System.IO.File.Delete(SebServiceLog);
 
 
             // Try to delete the "SebWindowsService" subdirectory
